Revert from using FLUSHALL to allow prefix usage
This reverts the change made in commit 4ee7c38a74fbd2d6dd8572d6e31f3b4bd4818afa, and adds proper handling for prefixes and versioning.

In that commit for the sake of speed, the code replaces a scan_iter("*") with client.flushdb().
The problem being that in Django, it's very common to use the PREFIX option to seperate caches, and to want to clear only one of the prefixed caches.
This actually didn't work in the previous code as well, as the hard-coded "*" would match all records regardless of prefix.

In this change, we replace the "*" with "{prefix}:*" to match all records for all versions in a cache but limited to the prefix for the cache itself.